fix(ts): Make changed properties passed into specific onChange handlers not optional
As an example, when `onInputValueChange` is called on a `useCombobox`, the `inputValue` state is known to have changed, and so the property in the change object should not be optional. This is already verified to be the case in tests.

For the `onStateChange` callbacks, all the properties are still optional because we don’t know what has changed.